Optimize ruletree node memory usage (#458)
* Optimize ruletree node memory usage

Reduce nodeKey struct size from 16 to 8 bytes by using a single type for kind
and tokenID. This saves memory and improves cache locality. I am making an
assumption that we will not have more than 2^4=16 kinds of nodes and not more than
2^28=268435456 unique tokens.

* Use single mutex to protect node data and children

Each node in ruletree have two RWMutex structures that is 48 bytes per node.
This change uses a single Mutex to protect both children and data
structures which uses only 8 bytes of memory. This is a trade-off
between concurrency and memory usage, but in practice the contention
is low and the memory savings are significant.

* Compact rule tree nodes to save memory

Once filter is done loading rules, we can compact the rule tree nodes to
save memory. Support is added to the filter only, server can still break
down rules into smaller chunks.